refactor(test): split compaction tests into their own module
Move the compact_session tests and their helper out of the oversized
config/tests.rs into config/compaction_tests.rs, and share EnvGuard via a
new config/test_support.rs. Reduces tests.rs lines-of-code, responsibilities,
and assertion-block count to improve CodeScene health; behavior unchanged.
